[logconfig] introduce config pkg for log files
this commit introduces a new package to load the configuration
for log files from environment and generalizes the previous implementation.
